agent/requirements.txt
langgraph==1.1.2
langchain==1.2.12
langchain-openai==1.1.11
langchain-community==0.4.1
langchain-chroma==1.1.0
chromadb==1.5.5
tavily-python==0.7.23
python-dotenv>=1.0.0
matplotlib>=3.7.0
